+++ src/extension.js
@@ -16,7 +16,7 @@ let syncer = null;
  */
 function activate(context) {
     console.log('Congratulations, your extension "sync-sftp" is now active!');
-
+    let isPaused = false
     let myStatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 100);
     let uploadSyncData = {}
     let reUploadSyncData = {}
@@ -49,14 +49,14 @@ function activate(context) {
             }
         );
     }
-    utils.updateStatusBarItem(myStatusBarItem, syncer)
-    statusBarInterval = setInterval(() => utils.updateStatusBarItem(myStatusBarItem, syncer), 1000)
+    const webviewProvider = utils.createWebViewProvider(context.extensionUri)
+    utils.updateStatusBarItem(myStatusBarItem, syncer, webviewProvider)
+    statusBarInterval = setInterval(() => utils.updateStatusBarItem(myStatusBarItem, syncer, webviewProvider), 1000)
     String.prototype.replaceAll = function (search, replacement) {
         const target = this;
         return target.replace(new RegExp(search, 'g'), replacement);
     };
 
-    const webviewProvider = utils.createWebViewProvider(context.extensionUri)
     context.subscriptions.push(
         vscode.window.registerWebviewViewProvider("sync-sftp.logView", webviewProvider)
     );
@@ -117,7 +117,6 @@ function activate(context) {
         }
     });
     const reconnect = vscode.commands.registerCommand('sync-sftp.reconnect', function () {
-        console.log('connect');
         syncer.connect()
     });
     const reUpload = vscode.commands.registerCommand('sync-sftp.reupload', function () {
@@ -167,6 +166,28 @@ function activate(context) {
             syncer.notifyAboutChanges()
         }
     });
+    const clearQuery = vscode.commands.registerCommand('sync-sftp.clearQuery', function () {
+        if (syncer) {
+            syncer.uploadFileFailed = [];
+            messenger.infoSuccess('Query cleared')
+        }
+    });
+    const toggleWatcher = vscode.commands.registerCommand('sync-sftp.toggleWatcher', function () {
+        isPaused = !isPaused
+
+        if (watcher && !watcher.isClosed()) {
+            watcher.close()
+        }
+        if (syncer) {
+            syncer.toggle()
+        }
+        if (!isPaused) {
+            runWatcher()
+        }
+
+        messenger.infoSuccess(`Sync SFTP is ${isPaused ? 'paused' : 'active'}`)
+    });
+    // toggleWatcher
 
     context.subscriptions.push(clear);
     context.subscriptions.push(reload);
@@ -175,6 +196,8 @@ function activate(context) {
     context.subscriptions.push(reUpload);
     context.subscriptions.push(makeEqual);
     context.subscriptions.push(detectDifferences);
+    context.subscriptions.push(clearQuery);
+    context.subscriptions.push(toggleWatcher);
     context.subscriptions.push(myStatusBarItem);
 }
 
@@ -186,7 +209,7 @@ function deactivate() {
     if (statusBarInterval) {
         clearInterval(statusBarInterval)
     }
-    if (syncer && syncer.timeInterval) {
+    if (syncer?.timeInterval) {
         clearInterval(syncer.timeInterval)
     }
 }